requirements.txt
numpy==2.0.0
pandas==2.2.2
python-dotenv>=0.5.1